2nd COMMIT
Cleaned up the main page a bit.
Added more souts to make it appear cleaner.
Added some string bits to each block of code for the "Answer:" +
Added a sout with a sting of  \n to space between the bloacks of code.
Lastly added some comments to describe what the blocks of code were doing.