Commit 3
We added 3 more functions to the JavaCalculator today.

In the Magic Calculator.java-

Added Reciprocals.

Added Exponents.

In the Calculator.java-

Added Percents %.

In the Main-

1.  Added Reciprocal and Exponent code blocks to the Magic Cal Class code block section.

2.  Added Percent to the  Calc Class Prints code block section.

3.  Finally followed my classmate Josh's idea about adding in and printing out num1 and num2 as numbers.
For example in Main.java, num1=4 and num2=2  so I added in the 3rd line of each code block, so it prints out  what its doing or what it is with num1 (4) or num2 (2).
ie Adding num1 and num2.

Also cleaned up more code and made it look more left aligned and added a welcome message at the top and the num 1 & 2's values.